fix nathan's problem

+++ src/ops/empty.rs
@@ -25,10 +25,12 @@ pub fn empty(dry: bool, days: Option<u32>) -> Result<()> {
                 println!("{:?}", file.path);
             } else {
                 fs::remove_file(file.info_path)?;
-                if file.deleted_path.is_dir() {
-                    fs::remove_dir_all(file.deleted_path)?;
-                } else {
-                    fs::remove_file(file.deleted_path)?;
+                if file.deleted_path.exists() {
+                    if file.deleted_path.is_dir() {
+                        fs::remove_dir_all(file.deleted_path)?;
+                    } else {
+                        fs::remove_file(file.deleted_path)?;
+                    }
                 }
             }
         }
